Author: BB Date: 06/01/2020
Updated Files
Added intents.json file for the bot to use to read from and construct 
responses from. Finished the learning part and implemented how the bot 
is able to search the json file and grab the best predicted match to 
respond with. Also updated the code to not have to train and run all of 
the ML related stuff each time it runs and instead is able to save it in 
a pickle file for later use.