Author: BB Date: 06/01/2020
upgraded to new version
upgraded to new tensorflow version as tf2.0 does not contain 
tensorflow.contrib and will give an error. However, after upgrading to 
the new version using tf-nightly it still does not run with the same 
error of not having contrib so I will probably not persue this further, 
either run with tf1.15 and python 3.6 or previous or do not worry about 
it.